requirements.txt

flask==3.0.0
numpy==1.26.2
joblib==1.3.2
werkzeug==3.0.1
scikit-learn==1.2.2
pygments==2.17.2
pillow==10.1.0
pandas==2.1.3
tensorflow==2.15.0
matplotlib==3.8.2

